#  opt: Add expression and xform rule for secondary index zigzag joins