--- a/pkg/sql/logictest/testdata/logic_test/pg_catalog
+++ b/pkg/sql/logictest/testdata/logic_test/pg_catalog
@@ -1334,9 +1334,9 @@ datestyle                          ISO, MDY      NULL      NULL        NULL     
 default_transaction_isolation      serializable  NULL      NULL        NULL        string
 default_transaction_read_only      off           NULL      NULL        NULL        string
 distsql                            2.0-off       NULL      NULL        NULL        string
+experimental_enable_zigzag_join    off           NULL      NULL        NULL        string
 experimental_force_lookup_join     off           NULL      NULL        NULL        string
 experimental_force_split_at        off           NULL      NULL        NULL        string
-experimental_force_zigzag_join     off           NULL      NULL        NULL        string
 experimental_serial_normalization  rowid         NULL      NULL        NULL        string
 experimental_vectorize             off           NULL      NULL        NULL        string
 extra_float_digits                 0             NULL      NULL        NULL        string
@@ -1378,9 +1378,9 @@ datestyle                          ISO, MDY      NULL  user     NULL      ISO, M
 default_transaction_isolation      serializable  NULL  user     NULL      default       default
 default_transaction_read_only      off           NULL  user     NULL      off           off
 distsql                            2.0-off       NULL  user     NULL      2.0-off       2.0-off
+experimental_enable_zigzag_join    off           NULL  user     NULL      off           off
 experimental_force_lookup_join     off           NULL  user     NULL      off           off
 experimental_force_split_at        off           NULL  user     NULL      off           off
-experimental_force_zigzag_join     off           NULL  user     NULL      off           off
 experimental_serial_normalization  rowid         NULL  user     NULL      rowid         rowid
 experimental_vectorize             off           NULL  user     NULL      off           off
 extra_float_digits                 0             NULL  user     NULL      0             2
@@ -1418,9 +1418,9 @@ datestyle                          NULL    NULL     NULL     NULL        NULL
 default_transaction_isolation      NULL    NULL     NULL     NULL        NULL
 default_transaction_read_only      NULL    NULL     NULL     NULL        NULL
 distsql                            NULL    NULL     NULL     NULL        NULL
+experimental_enable_zigzag_join    NULL    NULL     NULL     NULL        NULL
 experimental_force_lookup_join     NULL    NULL     NULL     NULL        NULL
 experimental_force_split_at        NULL    NULL     NULL     NULL        NULL
-experimental_force_zigzag_join     NULL    NULL     NULL     NULL        NULL
 experimental_serial_normalization  NULL    NULL     NULL     NULL        NULL
 experimental_vectorize             NULL    NULL     NULL     NULL        NULL
 extra_float_digits                 NULL    NULL     NULL     NULL        NULL