… pkg/sql/opt/bench/stub_factory.go
… pkg/sql/opt/memo/statistics_builder.go
… pkg/sql/opt/norm/testdata/rules/combo
… pkg/sql/opt/ops/relational.opt
… pkg/sql/opt/testutils/opt_tester.go
… pkg/sql/opt/xform/coster.go
… pkg/sql/opt/xform/custom_funcs.go
… pkg/sql/opt/xform/testdata/rules/join
… pkg/sql/opt/xform/testdata/rules/select
… pkg/sql/opt_exec_factory.go
… pkg/sql/plan_physical_props.go
… pkg/sql/sessiondata/session_data.go
… pkg/sql/vars.go
… pkg/sql/walk.go
… pkg/sql/zigzag_join.go