This change adds an optimizer expression, coster, statistics/logical
property builder as well as a simple xform rule to generate zigzag joins
in the case where we have select filters with const values on prefixes
of two separate secondary indexes.

Things to improve (either in this PR or in future ones): 
- [ ] Xform rules to catch more zigzaggable instances
- [x] More accurate statistics calculations (currently, zigzag joins wrapped with lookup joins can't compete with index-join selects due to the much higher row count)
- [ ] Better index selection algorithm instead of a quadratic iteration
- [ ] Refactor and split up `GenerateZigzagJoins` in `custom_funcs`  (probably best to do it alongside the above task)

Release note: None