@itsbilal 2018-11-20
Speedup with this change and this [zigzag-oriented benchmark](https://github.com/itsbilal/cockroach/blob/00ce4edf15a049292df08ab708bb54122d782066/pkg/bench/bench_test.go#L1130):

```
name                             old time/op  new time/op  delta
ZigzagJoin/Cockroach-4           12.0ms ±14%  10.3ms ±20%     ~     (p=0.056 n=5+5)
ZigzagJoin/CockroachOpt-4        11.1ms ±10%   0.7ms ±29%  -94.16%  (p=0.008 n=5+5)
ZigzagJoin/MultinodeCockroach-4  11.7ms ±15%   0.7ms ±34%  -94.20%  (p=0.008 n=5+5)
```