@rytaft 2018-11-27
___
*[pkg/sql/opt/memo/statistics_builder.go, line 306 at r9](https://reviewable.io/reviews/cockroachdb/cockroach/32512#-LSFlYTqDF0cKSOGm9rj:-LSM-Q388nhpE-6jzaTT:bd7yvcg) ([raw file](https://github.com/cockroachdb/cockroach/blob/0ff14498d821b1d3667531cfd3b4a4c872655389/pkg/sql/opt/memo/statistics_builder.go#L306)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

Now that I think about it, are you sure this code is actually being called? Logical props and stats are only calculated for normalized expressions, so if `ZigZagJoin` is always an alternate expression then it shouldn't need logical props or stats code. For example, `MergeJoin` does not have code for logical properties or stats. 

But if it is called, you can test the logical props by adding some test cases inside `memo/testdata/logprops`. You're right that the props show up in many other places, but it's a good idea to add targeted tests for the range of possible logical properties (see the other tests in that directory for examples).
</blockquote></details>

Sorry ignore that -- I see that it should be getting called since there is a ConstructZigZagJoin call inside GenerateZigZagJoins. But beware that it *won't* be called if the ZigZag is not the normalized expression in its group (e.g. if it's not wrapped in a lookup join).


<!-- Sent from Reviewable.io -->