@rytaft 2018-11-27
___
*[pkg/sql/opt/memo/testdata/stats/select, line 819 at r11](https://reviewable.io/reviews/cockroachdb/cockroach/32512#-LSM0BWE8XPL46bB25-B:-LSM0BWE8XPL46bB25-C:bq2rx3v) ([raw file](https://github.com/cockroachdb/cockroach/blob/823a6365d5c271ef12ed14d48de35cd0121ee6c5/pkg/sql/opt/memo/testdata/stats/select#L819)):*
> ```
> WHERE
>     l_shipdate = DATE '1995-09-01'
>     AND l_commitdate = DATE '1995-08-01';
> ```

I don't think this test is actually exercising your stats code. I think it's probably using the stats that were calculated for the normalized SELECT expression (which is probably why the stats are exactly the same as the norm case below...)


<!-- Sent from Reviewable.io -->