@rytaft 2018-11-27
___
*[pkg/sql/opt/memo/testdata/stats/select, line 819 at r11](https://reviewable.io/reviews/cockroachdb/cockroach/32512#-LSM0BWE8XPL46bB25-B:-LSMBX5H2taQnfUOzXfb:b-5b31xj) ([raw file](https://github.com/cockroachdb/cockroach/blob/823a6365d5c271ef12ed14d48de35cd0121ee6c5/pkg/sql/opt/memo/testdata/stats/select#L819)):*
<details><summary><i>Previously, itsbilal (Bilal Akhtar) wrote…</i></summary><blockquote>

It is - I was able to find a bug in `colStatJoin` this way. Though I don't think the single-column `colstat=11 colstat=12` directives are triggering colstat generation, but the `11,12` case definitely is.
</blockquote></details>

Ok, I guess those `colstat` directives are applied after optimization - makes sense. But I doubt this is testing the code in `buildZigzagJoin`. And it sounds like it's not testing the single-column stats. I'd suggest adding another test where the zigzag is wrapped in a lookup join so you test all the different code paths.


<!-- Sent from Reviewable.io -->