@rytaft 2018-11-27
___
*[pkg/sql/opt/memo/statistics_builder.go, line 1245 at r9](https://reviewable.io/reviews/cockroachdb/cockroach/32512#-LSFjg_98I_WWerBRf_W:-LSMCk4Y-ugSTlJ9w1os:b-igqhie) ([raw file](https://github.com/cockroachdb/cockroach/blob/0ff14498d821b1d3667531cfd3b4a4c872655389/pkg/sql/opt/memo/statistics_builder.go#L1245)):*
<details><summary><i>Previously, itsbilal (Bilal Akhtar) wrote…</i></summary><blockquote>

Done. Don't think a sanity check is necessary, actually. 
</blockquote></details>

I would make the comment more explicit (e.g., "the left and right indexes are from the same table, so the row count should be the same")


<!-- Sent from Reviewable.io -->